Refactored Module 3
