+++ .vscode/settings.json
@@ -5,5 +5,5 @@
     "python.testing.unittestEnabled": false,
     "python.testing.nosetestsEnabled": false,
     "python.testing.pytestEnabled": true,
-    "python.pythonPath": "venv/bin/python"
+    "python.pythonPath": ".venv/bin/python"
 }
