+++ pyproject.toml
@@ -10,6 +10,8 @@ flask = "^1.1.2"
 python-dotenv = "^0.14.0"
 requests = "^2.24.0"
 gunicorn = "^20.0.4"
+selenium = "^3.141.0"
+pytest = "^6.2.2"
 
 
 [tool.poetry.dev-dependencies]
